+++ tools/Azure.Mcp.Tools.Monitor/src/Commands/MonitorJsonContext.cs
@@ -6,16 +6,20 @@ using System.Text.Json.Serialization;
 using Azure.Mcp.Tools.Monitor.Commands.Metrics;
 using Azure.Mcp.Tools.Monitor.Commands.Table;
 using Azure.Mcp.Tools.Monitor.Commands.TableType;
+using Azure.Mcp.Tools.Monitor.Commands.WebTests;
 using Azure.Mcp.Tools.Monitor.Commands.Workspace;
 
 namespace Azure.Mcp.Tools.Monitor.Commands;
 
 [JsonSerializable(typeof(List<JsonNode>))]
-[JsonSerializable(typeof(WorkspaceListCommand.WorkspaceListCommandResult))]
+[JsonSerializable(typeof(MetricsDefinitionsCommand.MetricsDefinitionsCommandResult))]
+[JsonSerializable(typeof(MetricsQueryCommand.MetricsQueryCommandResult))]
 [JsonSerializable(typeof(TableListCommand.TableListCommandResult))]
 [JsonSerializable(typeof(TableTypeListCommand.TableTypeListCommandResult))]
-[JsonSerializable(typeof(MetricsQueryCommand.MetricsQueryCommandResult))]
-[JsonSerializable(typeof(MetricsDefinitionsCommand.MetricsDefinitionsCommandResult))]
+[JsonSerializable(typeof(WebTestsCreateOrUpdateCommand.WebTestsCreateCommandResult))]
+[JsonSerializable(typeof(WebTestsGetCommand.WebTestsGetCommandResult))]
+[JsonSerializable(typeof(WebTestsListCommand.WebTestsListCommandResult))]
+[JsonSerializable(typeof(WorkspaceListCommand.WorkspaceListCommandResult))]
 [JsonSourceGenerationOptions(PropertyNamingPolicy = JsonKnownNamingPolicy.CamelCase)]
 internal sealed partial class MonitorJsonContext : JsonSerializerContext
 {